pyproject.toml
[project]
name = "businessmanagment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "pydantic[email]>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "sqlalchemy-utils>=0.41.2",
    "uvicorn>=0.34.2",
    "fastapi-cache2[redis]>=0.2.2",
    "gunicorn>=23.0.0",
]
